Some objects were recreated each time 'paint()' was called, although each instance was exactly the same. I made class fields out of them that get assigned only one time, by the constructor.
Fields that are not supposed to change after they've been set by a constructor can be made final. This ensures that they are set (the compiler will warn you otherwise), and they cannot be changed afterwards.

Adding javadocs gives others a helpful hint in how to interpret this class. This makes future enhancements easier.

git-svn-id: https://svn.java.net/svn/worldclock-application~svn/trunk@17 b57468bd-8988-41ea-ecec-821d63db5d79